+++ backend/insert_csv.py
@@ -5,42 +5,16 @@ This module is used to insert csv file to the database
 import pandas as pd
 from datetime import datetime
 from models import * 
-from mapping_enums import class_mapping, college_mapping, concentration_mapping, role_mapping, hsr_mapping
 
 
 #### MAIN ####
 # Import csv file for students' data
 df = pd.read_csv(r'/home/khanh/Documents/capstone-hub/backend/M22_Capstone_descriptions.csv')
 
-#### Process NA values ####
-# NA_VALUE = "NONE"
-
-# print("BEFORE PROCESSING NAs")
-# print(f"Number of NA values: {df.isnull().sum()}")
-
-# df = df.fillna(NA_VALUE)
-
-# print("AFTER PROCESSING NAs")
-# print(f"Number of NA values: {df.isnull().sum()}")
-
-enum_cols = {
-    "role": role_mapping,
-    "primary_major": college_mapping,
-    "secondary_major": college_mapping,
-    "primary_concentration": concentration_mapping,
-    "secondary_concentration": concentration_mapping,
-    "minor": college_mapping,
-    "minor_concentration": concentration_mapping,
-    "hsr_review": hsr_mapping
-    }
-
 
 # Function to create lists of data to inject to Models
 """
 Take data and put to dictionary
-If column is enum:
-    Get equivalent enum data -> Put to tables
-
 """
 def create_list_of_data(data, data_cols, model_cols, default_cols=None, default_vals=None):
     """
@@ -63,19 +37,7 @@ def create_list_of_data(data, data_cols, model_cols, default_cols=None, default_
             model_col = model_cols[i]
             data_col = data_cols[i]
             row_value = row[data_col]
-            if model_col in enum_cols:
-                # Map values from data_col to enum_mapping
-                # Then put values from enum_mapping to the row data
-                mapper = enum_cols[model_col] # Name of the mapper
-                key = row_value # Value of the row aka key of the mapper
-                if key in mapper:
-                    value = mapper[key] # Value to insert aka value of the mapper
-                # elif key == NA_VALUE:
-                #     value = ""
-                else:
-                    value = key
-                row_data[model_col] = value
-            elif data_col == "timestamp":
+            if data_col == "timestamp":
                 datetime_object = datetime.strptime(row_value, '%m/%d/%Y %H:%M:%S')
                 row_data[model_col] = datetime_object
             else:
@@ -111,8 +73,8 @@ users_data_cols = ["id", "id", "name",\
                 "primary_major", "second_major",\
                 "primary_concentration", "second_concentration",\
                 "special_concentration", "minor", "minor_concentration"]
-users_default_cols = ["role", "class_year"]
-users_default_vals = [RoleEnum.student, ClassEnum.M22]
+users_default_cols = ["class_year"]
+users_default_vals = [2022]
 users_to_add = create_list_of_data(\
     data=df,\
     data_cols=users_data_cols,\
@@ -166,9 +128,9 @@ def insert_data():
                 stmt = table(**dict_row)
                 db.session.add(stmt)
                 db.session.commit()
-            except:
+            except Exception as e:
                 db.session.rollback()
-                print("""There's some error!""")
+                print(e)
                 continue
             finally:
                 db.session.close()
